--- src/main/java/uk/gov/companieshouse/acsp/notification/EmailNotification.java
@@ -1,4 +1,0 @@
-package uk.gov.companieshouse.acsp.notification;
-
-public interface EmailNotification {
-}
