+++ src/test/java/uk/gov/companieshouse/acsp/service/AcspServiceTest.java
@@ -238,7 +238,7 @@ class AcspServiceTest {
         var submissionData = new AcspDataSubmissionDao();
         Map<String, String> links = new HashMap<>();
         links.put("self", "/transaction/" + TRANSACTION_ID + "/");
-        submissionData.setLinks(links);
+        application.setLinks(links);
         application.setAcspDataSubmission(submissionData);
         ResponseEntity<Object> expectedResponse = new ResponseEntity<>(HttpStatus.NO_CONTENT);
         Transaction transaction = new Transaction();
@@ -256,7 +256,7 @@ class AcspServiceTest {
         var submissionData = new AcspDataSubmissionDao();
         Map<String, String> links = new HashMap<>();
         links.put("self", "/transaction/" + TRANSACTION_ID + "/");
-        submissionData.setLinks(links);
+        application.setLinks(links);
         application.setAcspDataSubmission(submissionData);
         ResponseEntity<Object> expectedResponse = new ResponseEntity<>(HttpStatus.NOT_FOUND);
         Transaction transaction = new Transaction();
@@ -280,7 +280,7 @@ class AcspServiceTest {
         var submissionData = new AcspDataSubmissionDao();
         Map<String, String> links = new HashMap<>();
         links.put("self", "/transaction/" + TRANSACTION_ID + "/");
-        submissionData.setLinks(links);
+        application.setLinks(links);
         application.setAcspDataSubmission(submissionData);
         ResponseEntity<Object> expectedResponse = new ResponseEntity<>(HttpStatus.FORBIDDEN);
         Transaction transaction = new Transaction();
@@ -303,7 +303,7 @@ class AcspServiceTest {
         var submissionData = new AcspDataSubmissionDao();
         Map<String, String> links = new HashMap<>();
         links.put("self", "/transaction/" + TRANSACTION_ID + "/");
-        submissionData.setLinks(links);
+        application.setLinks(links);
         application.setAcspDataSubmission(submissionData);
         ResponseEntity<Object> expectedResponse = new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
